* Fixed that NodeRating struct in app-server-rs had the wrong type for the value field. (needed f32, not f64) * Made progress attempting to implement the new layout mode (which creates fewer unnecessary "gaps" vertically), but then realized I need a dedicated solution to the problem. (since collisions can happens if you expand your chidlren upward, even if eg. you're the topmost entry among your siblings) I'm thus starting work on a library for it. (having it separate for a faster testing cycle, and because it might be useful to others eventually)

+++ Packages/app-server-rs/src/db/node_ratings.rs
@@ -14,7 +14,7 @@ pub struct NodeRating {
     r#type: String,
 	creator: String,
 	createdAt: i64,
-	value: f64,
+	value: f32,
 }
 impl From<tokio_postgres::row::Row> for NodeRating {
 	fn from(row: tokio_postgres::row::Row) -> Self {
